Minimize additional features for the `zip` dependency.
For tch, zip is used to read and write `.npz` archive files which use
Python's ZipFile and doesn't need any fancy feature: we can safely
assume that every file uses compression methods 0 (stored) or 8
(deflate), and encryption is never used. Only actual variations are
flate2's deflate backends, but any non-default backends do override
the default backend of minix-oxide so we don't necessarily forbid
the use of other backends as is.

For torch-sys, zip is used to extract downloaded libtorch archives,
which might be huge (up to 3 GB) but also don't need any fancy feature.
Alternative backends can still be faster to decompress them, but this
cost is one-off and miniz is not extremelly slower than others anyway.